🚀 Major ML Architecture Refactoring: ML Common Utilities & Fast Fail Validation
✨ **ML Common Utilities Suite:**
- Created comprehensive ml_common/ directory with 13 specialized utilities
- Added data_quality, pipeline_orchestrator, parallel_processing, memory_optimization
- Implemented model_evaluation, feature_selection, hpo_utils, cv_utils
- Added lookahead_protection, model_registry, base_safeguards
- Created validation_utils.py with enterprise-grade validation suite

🛡️ **Fast Fail Validation System:**
- **ConfigurationValidator**: Pre-initialization config validation with fast fail
- **DataValidator**: Comprehensive data integrity, schema, and statistical validation
- **ResourceValidator**: Memory/CPU monitoring with resource exhaustion prevention
- **ExecutionValidator**: Async timeout protection and result validation
- **ResultValidator**: Quality scoring and labeling result validation
- **MLValidationSuite**: Orchestrated validation pipeline with multi-stage error handling

🔧 **Step05 Streamlining & Cleanup:**
- Refactored step5_labeling.py to use centralized validation utilities
- Removed 15+ redundant step05_*.py files (800+ lines of duplicate code)
- Simplified from 800+ lines to clean, focused implementation
- Maintained all functionality while improving maintainability

🎯 **Architecture Benefits:**
- DRY principle: Single source of truth for validation logic
- Reusable: Any ML step can leverage the validation suite
- Consistent: Same fast fail behavior across all components
- Maintainable: Centralized updates apply everywhere
- Enterprise-grade: Comprehensive error classification and logging

📊 **Fast Fail Classifications:**
- Configuration issues → FAST FAIL
- Data corruption → FAST FAIL
- Resource exhaustion → FAST FAIL
- Timeouts → FAST FAIL
- Invalid results → FAST FAIL

Files changed: 50+ | Lines: ~2000+ | New utilities: 13

+++ src/training/steps/standardized_parquet_handler.py
@@ -58,7 +58,7 @@ class StandardizedParquetHandler:
             'exchange': 'string',
             'symbol': 'string',
             'timeframe': 'string',
-            'year': 'int16',
+            'year': 'int32',
             'month': 'int8',
             'day': 'int8',
             'trade_volume': 'float64',
@@ -161,7 +161,7 @@ class StandardizedParquetHandler:
                 for col in partition_cols:
                     if col in df.columns:
                         if col in ['year', 'month', 'day']:
-                            partition_fields.append(pa.field(col, pa.int16() if col == 'year' else pa.int8()))
+                            partition_fields.append(pa.field(col, pa.int32() if col == 'year' else pa.int8()))
                         else:
                             partition_fields.append(pa.field(col, pa.string()))
